step3 : refactor : switch => else if, MenuCount.js render함수 추가, MenuList.js 변수명 수정

+++ src/js/App.js
@@ -26,50 +26,43 @@ function App() {
             onCategoryChange: this.onCategoryChange,
             $nav: $('nav'),
         });
-        this.setState(await Api.getMenuList(this.currentCategoryName));
+        this.setState();
     };
 
-    this.setState = updatedItems => {
-        this.menuList.setState(updatedItems);
-        this.menuCount.setState(updatedItems.length);
+    this.setState = async () => {
+        const menuList = await Api.getMenuList(this.currentCategoryName);
+        this.menuList.setState(menuList);
+        this.menuCount.setState(menuList.length);
     };
 
     this.onAdd = async menuName => {
         await Api.postMenu(this.currentCategoryName, menuName);
-        this.setState(await Api.getMenuList(this.currentCategoryName));
+        this.setState();
     };
 
     this.onAction = (actionType, target, newMenuName) => {
         const id = target.closest('.menu-list-item').dataset.menuId;
         this.action(actionType, id, newMenuName);
+        this.setState();
     };
 
     this.onCategoryChange = async target => {
         if (target.classList.contains('cafe-category-name')) {
             $('#category-title').innerText = `${target.innerText} 메뉴 관리`;
             this.currentCategoryName = target.dataset.categoryName;
-            this.setState(await Api.getMenuList(this.currentCategoryName));
+            this.setState();
         }
     };
 
     this.action = async (actionType, id, newMenuName) => {
-        switch (actionType) {
-            case 'update': {
-                await Api.updateMenu(this.currentCategoryName, id, newMenuName);
-                break;
-            }
-            case 'delete': {
-                await Api.deleteMenu(this.currentCategoryName, id);
-                break;
-            }
-            case 'soldout': {
-                await Api.updateSoldout(this.currentCategoryName, id);
-                break;
-            }
+        if ('update' == actionType) {
+            await Api.updateMenu(this.currentCategoryName, id, newMenuName);
+        } else if ('delete' == actionType) {
+            await Api.deleteMenu(this.currentCategoryName, id);
+        } else if ('soldout' == actionType) {
+            await Api.updateSoldout(this.currentCategoryName, id);
         }
-        this.setState(await Api.getMenuList(this.currentCategoryName));
     };
-
 }
 
 const app = new App();
